refactor(agents): move the staged-item mappers into their agent packages
`claudeStagedItems` / `codexStagedItems` / `opencodeStagedItems` sat together in
`sandbox-core/src/sync/agent-propagate.ts` — three functions, one per agent, in a
layer whose whole job is to be agent-neutral. Each knows only its own agent's
pull shape (claude's category rels and the plugins/cache subpath, codex's
prompts-is-a-dir rule, opencode's `config/` prefix), so each moved to
`packages/agent-<id>/src/staged-items.ts` with its test.

The `StagedItem` contract, `stageItemsViaTransport` and the propagate machinery
stay shared: those are genuinely agent-neutral, and nothing about them changed.

The three `download-<agent>.ts` commands import from the agent packages directly
now instead of pulling these back through the `sandbox-core` barrel. Those
commands are themselves on the allowlist under phase 2, so this is one less
thing tying them to the shared layer when they move.

Allowlist: 14 -> 13.
